--- frontend/src/components/Sign/Generate/index.js
@@ -1,1 +1,0 @@
-export {default} from './Generate.vue'
