--- frontend/src/components/Sign/Profile/index.js
@@ -1,1 +1,0 @@
-export {default} from './profile.vue'
