--- frontend/src/components/Sign/Signed/index.js
@@ -1,1 +1,0 @@
-export {default} from './Upload.vue'
