--- frontend/src/components/Sign/Upload/index.js
@@ -1,1 +1,0 @@
-export {default} from './Upload.vue'
