+++ frontend/src/views/FillInRecipientPage.vue
@@ -8,7 +8,7 @@
 </template>
 
 <script>
-import BadgeClassCard from "@/components/BadgeClassCard";
+import BadgeClassCard from "Components/BadgeClassCard";
 
 export default {
   components: {
