+++ frontend/src/views/ShareBadgeLinkPage.vue
@@ -9,7 +9,7 @@
 </template>
 
 <script>
-import BadgeClassCard from "@/components/BadgeClassCard";
+import BadgeClassCard from "Components/BadgeClassCard";
 
 export default {
   data() {
